stat: Add flags for statistic logging classes and correct function calls.
class 1. RTT statistics logging, as default log class
class 2. One way logging, used mainly for one way latency.

On the mirror we plan to add the one-way latency, so adding the correct flag
On the reference, the one way and RTT will be present, for one way for
frames coming from mirror and RTT for the usual application.

Signed-off-by: Walfred Tedeschi <walfred.tedeschi@intel.com>